refactor: 2024-12-30 고정날짜로 수정
