+++ src/main/java/attendance/CurrentDate.java
@@ -1,0 +1,7 @@
+package attendance;
+
+import java.time.LocalDate;
+
+public interface CurrentDate {
+    LocalDate now();
+}
