+++ src/main/java/attendance/SystemDate.java
@@ -1,0 +1,11 @@
+package attendance;
+
+import java.time.LocalDate;
+
+public class SystemDate implements CurrentDate{
+
+    @Override
+    public LocalDate now() {
+        return LocalDate.of(2024, 12, 30);
+    }
+}
